Refactor e2e test helpers into shared lib and use dynamic validator keys
Move docker compose helpers (exec_ok, dump_compose_logs,
start_compose, create_extra_wallet, etc.) and shared constants from
committee_rotation.rs into lib.rs so they can be reused by future e2e tests.

Replace hardcoded test validator keys with dynamically generated ones
using ValidatorKeypair::generate_from, avoiding issues with
well-known keys.

Replace hardcoded light client contract address with derived one.
